fixed ts

+++ library-UI/library-app/src/app/books/books.component.ts
@@ -110,8 +110,8 @@ export class BooksComponent implements OnInit {
     }
   }
 
-  deleteBook(id: number){
-    if(window.confirm('Сигурни ли сте, че искате да изтриете книгата?')) {
+  deleteBook(id: number,currentNumbers:number){
+    if(currentNumbers&&window.confirm('Сигурни ли сте, че искате да изтриете книгата?')) {
       this.bookService.deleteBookByID(id).subscribe(
         () => {
           this.reloadData();
